+++ packages/react/src/contexts/socketContext.tsx
@@ -1,5 +1,5 @@
 import log from '@kengoldfarb/log'
-import { MeemAPI } from '@meemproject/api'
+import { MeemAPI } from '@meemproject/sdk'
 import { isEqual, remove } from 'lodash'
 import React, { useContext, createContext, useState, ReactNode } from 'react'
 import { MatchMutate, useMatchMutate } from '../lib/useMatchMutate'