+++ packages/react/src/index.tsx
@@ -1,7 +1,19 @@
+// Contexts
 export * from './contexts/socketContext'
-export * from './contexts/walletContext'
+export * from './contexts/meemContext'
+export * from './contexts/authContext'
+export * from './contexts/meemUserContext'
+
+// Lib / utilities
 export * from './lib/getNodeSize'
 export * from './lib/useMatchMutate'
 export * from './lib/useStateCallback'
-export * from '../src/components/ExampleComponent'
 export * from './lib/chains'
+
+// GQL
+export * from '../generated/graphql'
+export * from './gql/auth'
+
+// Components
+export * from './components/LoginModal'
+export * from '../src/components/ExampleComponent'