+++ packages/sdk/src/methods/mint.ts
@@ -1,0 +1,3 @@
+// import {} from '@meemproject/meem-contracts'
+
+export const mint = () => {}